Windows build

+++ src/app/Application.cpp
@@ -14,7 +14,7 @@ Application::Application():
 
     const auto fullscreenFlag = config_.get<bool>(Setting::Fullscreen) ? sf::Style::Fullscreen : 0;
 
-    window_.create(sf::VideoMode{config_.get<int>(Setting::ResWidth), config_.get<int>(Setting::ResHeight)},
+    window_.create(sf::VideoMode{static_cast<unsigned>(config_.get<int>(Setting::ResWidth)), static_cast<unsigned>(config_.get<int>(Setting::ResHeight))},
         PROJECT_NAME, sf::Style::Titlebar | sf::Style::Close | fullscreenFlag, windowSettings);
     window_.setVerticalSyncEnabled(true);
